+++ src/opportunities/components/AboutSectionElement.jsx
