+++ src/staff/components/LargeImageCard.jsx
